Commented 2 functions I never used just to get rid of warning (this whole branch will be trashed at some point anyway if we switch to the swift version)

+++ DDMathParser/_DDFunctionEvaluator.m
@@ -417,9 +417,9 @@ static NSString *const _DDFunctionSelectorSuffix = @":variables:error:";
         upperBound = [[params objectAtIndex:1] longLongValue];
     }
     
-    long long random = arc4random_uniform(upperBound);
+    long long random = arc4random_uniform((uint32_t)upperBound);
     while (random < lowerBound) {
-        random = arc4random_uniform(upperBound);
+        random = arc4random_uniform((uint32_t)upperBound);
     }
 	
 	return [DDExpression numberExpressionWithNumber:@(random)];
